# feat(ext/http): stream auto resp body compression