--- a/ext/http/lib.rs
+++ b/ext/http/lib.rs
@@ -707,7 +707,12 @@ async fn op_http_write(
       }
     };
 
-    match body_writer.write_all(&buf).await {
+    let mut res = body_writer.write_all(&buf).await;
+    if res.is_ok() {
+      res = body_writer.flush().await;
+    }
+
+    match res {
       Ok(_) => break Ok(()),
       Err(err) => {
         assert_eq!(err.kind(), std::io::ErrorKind::BrokenPipe);
@@ -734,7 +739,21 @@ async fn op_http_shutdown(
     .resource_table
     .get::<HttpStreamResource>(rid)?;
   let mut wr = RcRef::map(&stream, |r| &r.wr).borrow_mut().await;
-  take(&mut *wr);
+  let wr = take(&mut *wr);
+  match wr {
+    HttpResponseWriter::Body(mut body_writer) => {
+      match body_writer.shutdown().await {
+        Ok(_) => {}
+        Err(err) => {
+          assert_eq!(err.kind(), std::io::ErrorKind::BrokenPipe);
+          // Don't return "broken pipe", that's an implementation detail.
+          // Pull up the failure associated with the transport connection instead.
+          stream.conn.closed().await?;
+        }
+      }
+    }
+    _ => {}
+  }
   Ok(())
 }