This commit adds support for auto response body compression for
streaming bodies.

Alternative to https://github.com/denoland/deno/tree/compression_stream
that uses `async_compression` and tokio's io traits.

cc @AaronO @satyarohith 

Fixes #13835