@lucacasonato 2022-04-20
@satyarohith Added the tests, and fixed two bugs in the process.