@littledivy 2022-04-20
How bad does this affect performance? Benchmark results of https://github.com/denoland/deno/pull/14328/files will be useful